Move PolledExecutorFacade status into BlockProviderExecutor
This makes this status available to the scale_in methods of
BlockProviderExecutors, and so can be used as part of a scaling bugfix
in a subsequent PR.

This is part of ongoing work to minimise and possibly eliminate the
PolledExecutorFacade class.

This PR leaves accesses to those structures in the PolledExecutorFacade,
with future PRs intended to evaporate away more of the PolledExecutorFacade.

This should not change behaviour, as it only moves an attribute from one
object to a 1:1 paired object.