compute monitoring deltas on each poll
this means this monitoring code does not need to know if a cache-refresh happened or not

subtleties: if someone modifies the poller dict elsewhere, this won't catch that becuase its sharing that mutable old state
and so that mutator needs to send a monitoring notification themselves.

that's how things are already i think: pollitem.scale_out and pollitem.scale_in already do that monitoring change

this is probably broken - get some better testing. or perhaps redo how monitoring deltas happen entirely?