Exit monitoring router exit on multiprocessing event, not exit message
Prior to this PR, the monitoring router exited due to receiving a
WORKFLOW_INFO message with an exit_now field set to True, but only if that
message was received through a specific path.

This PR removes that exit_now field, and makes the monitoring router exit
on a multiprocessing event. This removes the need for the exit message to
arrive through that specific path into the router, which makes message
handling more consistent, and opens up opportunities to feed messages into
monitoring through different paths.

Slowly ongoing work has been trying to make all the different monitoring
message paths behave the same with a goal of eliminating some of them, and
this change also works towards that.